requirements.txt
amqp==2.4.2
billiard==3.6.0.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.3
entrypoints==0.3
flake8==3.7.7
flake8-commas==2.0.0
flake8-quotes==2.0.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
kombu==4.5.0
mccabe==0.6.1
oauthlib==3.0.1
psycopg2-binary==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.2
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
